reporting, and fixes to actions
